hot fix: query uniswap exchange rate of latest block
For BSC & Polygon, the node seem to only keep track of states
in the last 1 block (?!) so that when we CallContract the
bn we just queries is already discarded.

Fix: re-order the BlockNumber() and CallContract() functions
so that we always get the state; however the blockNum we
attach to it might not be very accurate.