hot fix: exclude inBlockHeight from sendHash calculation
We had a strange issue that the same inTx on Ropsten
0x024196a9577b9ca5c66b8e43b2aea2bf29c6168b2daabc677e3b3217d2900be3
registers on ZetaChain as two CCTX; the reason is that
on June 9 our Ropsten node says the tx above is in
block 12352858, whereas on June 10 our Ropsten node
says the same tx is in block 12359086.

Because we hash the inBlockHeight, this inTx is treated
as two separate txs which is wrong.