zevm: cap depositAndCall() gas limit to 1M
since right now we do not have a mechanism for users to pay
for zEVM gas if they invoke zEVM contracts from other chains
via the DepositAndCall(). The finalizing validator vote msg
basically subsidizes the gas cost for the user, and for now
we use a simple hard cap to avoid abuse.